+++ server/src/main/java/org/apache/iotdb/db/metadata/MManager.java
@@ -805,12 +805,8 @@ public class MManager {
   }
 
   /** To calculate the count of timeseries for given prefix path. */
-  public int getAllTimeseriesCount(PartialPath prefixPath) throws MetadataException {
-    //    if (prefixPath.getNodes().length == 1) {
-    //      return (int) totalSeriesNumber.get();
-    //    }
-    //    return mtree.getAllTimeseriesCount(prefixPath);
-    return (int) totalSeriesNumber.get();
+  public long getAllTimeseriesCount(PartialPath prefixPath) throws MetadataException {
+    return totalSeriesNumber.get();
   }
 
   /** To calculate the count of devices for given prefix path. */
